Additional keys for com.github.spotbugs and related dependencies
This is to resolve the build failure of 85ae9af16f79b951b262164e9869f5dbf0d0e731

-- com.github.spotbugs --

Signature resolves to "Jeremy Landis (hazendaz) <jeremylandis@hotmail.com>".

This matches the GitHub identity "hazendaz" of the commit associated with the most recent release:
https://github.com/spotbugs/spotbugs-maven-plugin/commit/741f662a0d32ca1d4307d5bf47f75424893104ff

The commits are not signed, however, so we cannot compare any specific PGP keys.

-- com.github.javaparser dependency --

Signature resolves to "Roger Howell <MysterAitch@users.noreply.github.com>".

This matches the GitHub identity "MysterAitch" of the commit associated with the most recent release:
https://github.com/javaparser/javaparser/commit/67d7edf70b7cb03a2c717b74b5ed23c640de22f3

The commits are not signed, however, so we cannot compare any specific PGP keys.