new key for org.codehaus.mojo:versions-maven-plugin
Signature resolves to "Stefan Seifert <stefan.seifert@diva-e.com>"

GitHub user "stefanseifert" is the committer on the release's associate tag:
   https://github.com/mojohaus/versions-maven-plugin/releases
   https://github.com/mojohaus/versions-maven-plugin/tree/versions-maven-plugin-2.10.0
   https://github.com/stefanseifert

None of the GitHub commits appear to be signed, however, so cannot specifically validate this PGP key.

The identity, as much as we can verify, does look correct for this project.