new keys for cglib:cglib
The oldest artifacts are unsigned.

The oldest signature resolves to "Maven Central Java.net import".  This key is already in the map for several other projects.

The next oldest artifacts are signed by "Stuart McCulloch (Sonatype) <mcculls@sonatype.com>".  This key is also already in the map for "org.sonatype.*"

The newest artfacts are signed by "Sam Berlin <sberlin@gmail.com>".  It is worth noting this matches the GitHub user of the recent releases:
https://github.com/cglib/cglib/releases/tag/RELEASE_3_3_0
https://github.com/sameb